Checkpoint work to find stutter in fseq playback Fixed crash when playing remote files

+++ include/input/InputFPPRemotePlayFile.hpp
@@ -81,7 +81,7 @@ private:
     void        UpdateElapsedPlayTimeMS ();
     uint32_t    CalculateFrameId (uint32_t ElapsedMS, int32_t SyncOffsetMS);
     bool        ParseFseqFile ();
-    uint32_t    ReadFile(uint32_t DestinationIntensityId, uint32_t NumBytesToRead, uint32_t FileOffset);
+    uint64_t    ReadFile(uint64_t DestinationIntensityId, uint64_t NumBytesToRead, uint64_t FileOffset);
 
     String      LastFailedPlayStatusMsg;
